pyproject.toml
[project]
name = "mini-challenge-msr-2026"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "gitpython (>=3.1.45,<4.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)",
    "fastparquet (>=2024.11.0,<2025.0.0)",
    "huggingface-hub (>=1.1.5,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scipy (>=1.16.3,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
